# src/swagger.yaml
swagger: "2.0"
info:
  description: "Video Sharing Platform"
  version: "1.0.0"
  title: "Video Sharing Platform API"
  termsOfService: "http://swagger.io/terms/"
  contact:
    email: "developersworkspace@gmail.com"
  license:
    name: "MIT"
    url: "https://opensource.org/licenses/MIT"
host: "localhost:3000"
basePath: "/api"
schemes:
- "http"
paths:
  /profile:
    get:
      summary: "Find Profile"
      description: "Returns authenticated or specified profile"
      operationId: "findProfile"
      produces:
      - "application/json"
      parameters:
      - in: "query"
        name: "name"
        description: "Profile Name"
        required: true
        type: "string"
      responses:
        200:
            description: "Profile"
            schema:
                $ref: "#/definitions/Profile"
    put:
      summary: "Update Profile"
      description: "Updates profile"
      operationId: "updateProfile"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "name"
        description: "Profile Name"
        required: true
        schema:
          $ref: "#/definitions/Profile"
      responses:
        200:
            description: "Profile"
            schema:
                $ref: "#/definitions/Profile"
definitions:
  Profile:
    type: "object"
    properties:
      address:
        $ref: "#/definitions/Address"
      contactDetails:
        $ref: "#/definitions/Address"
      description:
        type: "string"
      id:
        type: "string"
      message:
        type: "string"
      name:
        type: "string"
      socialDetails:
        $ref: "#/definitions/SocialDetails"
      userId:
        type: "string"
  Address:
    type: "object"
    properties:
      city:
        type: "string"
      country:
        type: "string"
  ContactDetails:
    type: "object"
    properties:
      emailAddress:
        type: "string"
      phoneNumber:
        type: "string"
  SocialDetails:
    type: "object"
    properties:
      facebook:
        type: "string"
      instagram:
        type: "string"
      twitter:
        type: "string"
externalDocs:
  description: "Find out more about Swagger"
  url: "http://swagger.io"